requirements.txt
Flask
python-dotenv
yagmail
cryptography